application/requirements.txt
asgiref==3.5.1
Django==4.0.6
pip==22.1
setuptools==58.1.0
sqlparse==0.4.2
tzdata==2022.1
django-environ==0.8.1
psycopg2-binary==2.9.3
gunicorn==20.1.0
sqlparse==0.4.2
